feat: introduce a crate for builder functionality
