+++ xcm-primitives/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "t3rn-xcm-primitives"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "3.1.5", default-features = false, features = ['derive'] }
+xcm = { git = "https://github.com/paritytech/polkadot.git", branch = "release-v0.9.27", default-features = false }
+
+[features]
+default = [ "std" ]
+std = [
+    "xcm/std",
+    "codec/std",
+]
